requirements.txt
python-telegram-bot==21.9
aiohttp>=3.9
rapidfuzz==3.6.1
python-dotenv>=1.0
